Update README.md
This is data on Lehigh University's website that are general to all US college students. The key findings are that statistics found there is a lack of reporting when sexual assault occurs on collehe campuses, and that women are more likely to be victims then men. This is a topic that needs to have awareness spread because it affects every college student.